requirements.txt
Django>=4.2
mongoengine>=0.27.0
djangorestframework>=3.14.0
ortools>=9.8.3296
